backend/requirements.txt
azure-core==1.29.5
azure-identity==1.15.0
azure-search-documents==11.4.0b8
Flask==3.0.0
Flask-Cors==4.0.0
gunicorn==21.2.0
openai==1.3.7
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==3.0.1
click==8.1.7
setuptools==69.0.2
wheel==0.42.0
